// Kerekparbolt/WebUI/src/services/client/bicycleService.ts
export interface BicycleResponse {
    Ar?: number | undefined;
    Cikkszam?: string | undefined;
    FelniAtmero?: string | undefined;
    FelniAtmeroID?: number | undefined;
    Id?: number | undefined;
    Marka?: string | undefined;
    MarkaID?: number | undefined;
    Tipus?: string | undefined;
    TipusID?: number | undefined;
    URL?: string | undefined;
    Valto?: string | undefined;
    ValtoTipus?: number | undefined;
    Vazmeret?: string | undefined;
    VazmeretID?: number | undefined;
}